@@ -8,19 +8,15 @@
     "webpack": "webpack --watch"
   },
   "devDependencies": {
-    "babel-core": "^6.18.0",
-    "babel-loader": "^6.2.7",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-register": "^6.18.0",
     "chai": "^3.5.0",
     "chai-as-promised": "^5.3.0",
     "chai-shallow-deep-equal": "^1.4.0",
-    "coveralls": "^2.11.14",
     "json-loader": "^0.5.4",
     "mocha": "^3.0.2",
     "mocha-lcov-reporter": "^1.2.0",
-    "source-map-support": "^0.4.10",
-    "webpack": "^1.14.0",
-    "yargs": "5.0.0"
+    "ts-loader": "^3.1.1",
+    "ts-node": "^3.3.0",
+    "typescript": "^2.6.1",
+    "webpack": "^3.8.1"
   }
 }
